**v.0.5.5 2017-05-23**
* `tidy_genomic_data` : bug fix introduce with previous commit when fixing
`LOCUS` and `COL` with stacks version > 1.44.
Thanks to Eric Archer for highlighting the bug.

* `summary_haplotypes`: this function gets a new arguments, `keep.consensus`, to
enable the calculation of **pi** to include or not the consensus markers presents
in stacks haplotype file (e.g. *batch_1.haplotypes.tsv*). This argument works to
circumvent the impact of using a whitelist of markers,
that potentially removed those markers in previous versions. Also changed in this
function, the summary table include a **POLYMORPHISM** column that no longer
include the artifact marker counts (markers with more than 2 alleles).
This information is kept in a separate column (as before).